Extract SleepActivityUITest from MainActivityUITest
Also add a UITestBase base class to avoid duplication.
